requirements.txt
requests>=2.28.0
web3>=6.0.0
secretsharing>=0.2.7
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
py-solc-x>=1.1.1
eth-tester>=0.9.0b2

